+++ release/flashlight/R/plot_light_ice.R
@@ -26,7 +26,7 @@
 #' indices <- (1:15) * 10
 #' plot(light_ice(mod_full, v = "Species"), rotate_x = TRUE)
 #' plot(light_ice(mods, v = "Species", indices = indices))
-#' plot(light_ice(mods, v = "Species", indices = indices, center = TRUE))
+#' plot(light_ice(mods, v = "Species", indices = indices, center = "first"))
 #' plot(light_ice(mods, v = "Petal.Width", n_bins = 5, indices = indices))
 #' plot(light_ice(mods, v = "Petal.Width", by = "Species", n_bins = 5, indices = indices))
 #'
@@ -40,8 +40,8 @@
 #'   predict_function = function(m, X) predict(m, X, type = "response"))
 #' fls <- multiflashlight(list(fl_lm, fl_glm), data = ir)
 #' plot(light_ice(fls, v = "Petal.Length", indices = indices))
-#' plot(light_ice(fls, v = "Petal.Length", indices = indices, center = TRUE))
-#' plot(light_ice(fls, v = "Petal.Length", indices = indices, by = "Species", center = TRUE))
+#' plot(light_ice(fls, v = "Petal.Length", indices = indices, center = "first"))
+#' plot(light_ice(fls, v = "Petal.Length", indices = indices, by = "Species", center = "first"))
 #' plot(light_ice(fls, v = "Petal.Length", indices = indices, use_linkinv = FALSE))
 #' @seealso \code{\link{light_ice}}.
 plot.light_ice <- function(x, facet_scales = "fixed", rotate_x = FALSE, ...) {